WIP: Change "product" to be a uint32_t in two places in tic.h.
Those functions have not been part of a release yet.

+++ include/tic.h
@@ -101,13 +101,13 @@ void tic_string_free(char *);
 /// product argument should be one of the TIC_PRODUCT_* macros, but if it is
 /// not, this function returns an empty string.
 TIC_API
-const char * tic_look_up_product_name_short(uint32_t product);
+const char * tic_look_up_product_name_short(uint8_t product);
 
 /// Looks up a user-friendly string for the product name.  The product argument
 /// should be one of the TIC_PRODUCT_* macros, but if it is not, this function
 /// returns "(Unknown)".
 TIC_API
-const char * tic_look_up_product_name_ui(uint32_t product);
+const char * tic_look_up_product_name_ui(uint8_t product);
 
 /// Looks up a user-friendly string corresponding to the specified error bit,
 /// e.g. "Safe start violation".  The error argument should be of the form (1 <<
